+++ pycoreloop/core_loop.py
@@ -1,7 +1,7 @@
 r"""Wrapper for core_loop.h
 
 Generated with:
-/home/anze/anaconda3/bin/ctypesgen ../coreloop/core_loop.h
+/u/home/anze/anaconda3/bin/ctypesgen ../coreloop/core_loop.h
 
 Do not modify this file.
 """
@@ -863,23 +863,23 @@ add_library_search_dirs([])
 
 # No modules
 
-__uint8_t = c_ubyte# /usr/include/x86_64-linux-gnu/bits/types.h: 38
+__uint8_t = c_ubyte# /usr/include/bits/types.h: 37
 
-__uint16_t = c_ushort# /usr/include/x86_64-linux-gnu/bits/types.h: 40
+__uint16_t = c_ushort# /usr/include/bits/types.h: 39
 
-__uint32_t = c_uint# /usr/include/x86_64-linux-gnu/bits/types.h: 42
+__uint32_t = c_uint# /usr/include/bits/types.h: 41
 
-__uint64_t = c_ulong# /usr/include/x86_64-linux-gnu/bits/types.h: 45
+__uint64_t = c_ulong# /usr/include/bits/types.h: 44
 
-uint8_t = __uint8_t# /usr/include/x86_64-linux-gnu/bits/stdint-uintn.h: 24
+uint8_t = __uint8_t# /usr/include/bits/stdint-uintn.h: 24
 
-uint16_t = __uint16_t# /usr/include/x86_64-linux-gnu/bits/stdint-uintn.h: 25
+uint16_t = __uint16_t# /usr/include/bits/stdint-uintn.h: 25
 
-uint32_t = __uint32_t# /usr/include/x86_64-linux-gnu/bits/stdint-uintn.h: 26
+uint32_t = __uint32_t# /usr/include/bits/stdint-uintn.h: 26
 
-uint64_t = __uint64_t# /usr/include/x86_64-linux-gnu/bits/stdint-uintn.h: 27
+uint64_t = __uint64_t# /usr/include/bits/stdint-uintn.h: 27
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/spectrometer_interface.h: 46
+# /fast/lusee/coreloop/coreloop/spectrometer_interface.h: 46
 class struct_ADC_stat(Structure):
     pass
 
@@ -903,11 +903,11 @@ struct_ADC_stat._fields_ = [
     ('sumv2', uint64_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 134
+# /fast/lusee/coreloop/coreloop/core_loop.h: 138
 class struct_core_state(Structure):
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 32
+# /fast/lusee/coreloop/coreloop/core_loop.h: 34
 for _lib in _libs.values():
     try:
         state = (struct_core_state).in_dll(_lib, "state")
@@ -915,7 +915,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 33
+# /fast/lusee/coreloop/coreloop/core_loop.h: 35
 for _lib in _libs.values():
     try:
         avg_counter = (uint16_t).in_dll(_lib, "avg_counter")
@@ -923,7 +923,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 34
+# /fast/lusee/coreloop/coreloop/core_loop.h: 36
 for _lib in _libs.values():
     try:
         unique_packet_id = (uint32_t).in_dll(_lib, "unique_packet_id")
@@ -931,7 +931,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 35
+# /fast/lusee/coreloop/coreloop/core_loop.h: 37
 for _lib in _libs.values():
     try:
         leading_zeros_min = (uint8_t * int(16)).in_dll(_lib, "leading_zeros_min")
@@ -939,7 +939,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 36
+# /fast/lusee/coreloop/coreloop/core_loop.h: 38
 for _lib in _libs.values():
     try:
         leading_zeros_max = (uint8_t * int(16)).in_dll(_lib, "leading_zeros_max")
@@ -947,7 +947,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 37
+# /fast/lusee/coreloop/coreloop/core_loop.h: 39
 for _lib in _libs.values():
     try:
         housekeeping_request = (uint8_t).in_dll(_lib, "housekeeping_request")
@@ -955,7 +955,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 38
+# /fast/lusee/coreloop/coreloop/core_loop.h: 40
 for _lib in _libs.values():
     try:
         section_break = (uint32_t).in_dll(_lib, "section_break")
@@ -963,7 +963,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 39
+# /fast/lusee/coreloop/coreloop/core_loop.h: 41
 for _lib in _libs.values():
     try:
         range_adc = (uint8_t).in_dll(_lib, "range_adc")
@@ -971,7 +971,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 39
+# /fast/lusee/coreloop/coreloop/core_loop.h: 41
 for _lib in _libs.values():
     try:
         resettle = (uint8_t).in_dll(_lib, "resettle")
@@ -979,7 +979,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 39
+# /fast/lusee/coreloop/coreloop/core_loop.h: 41
 for _lib in _libs.values():
     try:
         request_waveform = (uint8_t).in_dll(_lib, "request_waveform")
@@ -987,7 +987,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 40
+# /fast/lusee/coreloop/coreloop/core_loop.h: 42
 for _lib in _libs.values():
     try:
         tick_tock = (c_bool).in_dll(_lib, "tick_tock")
@@ -995,7 +995,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 41
+# /fast/lusee/coreloop/coreloop/core_loop.h: 43
 for _lib in _libs.values():
     try:
         drop_df = (c_bool).in_dll(_lib, "drop_df")
@@ -1003,7 +1003,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 42
+# /fast/lusee/coreloop/coreloop/core_loop.h: 44
 for _lib in _libs.values():
     try:
         soft_reset_flag = (c_bool).in_dll(_lib, "soft_reset_flag")
@@ -1011,7 +1011,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 43
+# /fast/lusee/coreloop/coreloop/core_loop.h: 45
 for _lib in _libs.values():
     try:
         heartbeat_packet_count = (uint32_t).in_dll(_lib, "heartbeat_packet_count")
@@ -1019,39 +1019,47 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 44
+# /fast/lusee/coreloop/coreloop/core_loop.h: 46
 for _lib in _libs.values():
     try:
-        heartbeat_counter = (uint32_t).in_dll(_lib, "heartbeat_counter")
+        heartbeat_counter = (uint64_t).in_dll(_lib, "heartbeat_counter")
         break
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 45
+# /fast/lusee/coreloop/coreloop/core_loop.h: 47
 for _lib in _libs.values():
     try:
-        resettle_counter = (uint32_t).in_dll(_lib, "resettle_counter")
+        resettle_counter = (uint64_t).in_dll(_lib, "resettle_counter")
         break
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 46
+# /fast/lusee/coreloop/coreloop/core_loop.h: 48
 for _lib in _libs.values():
     try:
-        cdi_wait_counter = (uint32_t).in_dll(_lib, "cdi_wait_counter")
+        cdi_wait_counter = (uint64_t).in_dll(_lib, "cdi_wait_counter")
         break
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 47
+# /fast/lusee/coreloop/coreloop/core_loop.h: 49
 for _lib in _libs.values():
     try:
-        mini_wait_counter = (uint32_t).in_dll(_lib, "mini_wait_counter")
+        cdi_dispatch_counter = (uint64_t).in_dll(_lib, "cdi_dispatch_counter")
         break
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 48
+# /fast/lusee/coreloop/coreloop/core_loop.h: 50
+for _lib in _libs.values():
+    try:
+        tap_counter = (uint64_t).in_dll(_lib, "tap_counter")
+        break
+    except:
+        pass
+
+# /fast/lusee/coreloop/coreloop/core_loop.h: 51
 for _lib in _libs.values():
     try:
         flash_store_pointer = (uint16_t).in_dll(_lib, "flash_store_pointer")
@@ -1059,27 +1067,27 @@ for _lib in _libs.values():
     except:
         pass
 
-enum_gain_state = c_int# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+enum_gain_state = c_int# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-GAIN_LOW = 0# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+GAIN_LOW = 0# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-GAIN_MED = (GAIN_LOW + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+GAIN_MED = (GAIN_LOW + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-GAIN_HIGH = (GAIN_MED + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+GAIN_HIGH = (GAIN_MED + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-GAIN_DISABLE = (GAIN_HIGH + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+GAIN_DISABLE = (GAIN_HIGH + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-GAIN_AUTO = (GAIN_DISABLE + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 53
+GAIN_AUTO = (GAIN_DISABLE + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 56
 
-enum_output_format = c_int# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 60
+enum_output_format = c_int# /fast/lusee/coreloop/coreloop/core_loop.h: 63
 
-OUTPUT_32BIT = 0# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 60
+OUTPUT_32BIT = 0# /fast/lusee/coreloop/coreloop/core_loop.h: 63
 
-OUTPUT_16BIT_UPDATES = (OUTPUT_32BIT + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 60
+OUTPUT_16BIT_UPDATES = (OUTPUT_32BIT + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 63
 
-OUTPUT_16BIT_FLOAT1 = (OUTPUT_16BIT_UPDATES + 1)# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 60
+OUTPUT_16BIT_FLOAT1 = (OUTPUT_16BIT_UPDATES + 1)# /fast/lusee/coreloop/coreloop/core_loop.h: 63
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 66
+# /fast/lusee/coreloop/coreloop/core_loop.h: 69
 class struct_route_state(Structure):
     pass
 
@@ -1093,7 +1101,7 @@ struct_route_state._fields_ = [
     ('minus', uint8_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 72
+# /fast/lusee/coreloop/coreloop/core_loop.h: 75
 class struct_sequencer_state(Structure):
     pass
 
@@ -1139,7 +1147,7 @@ struct_sequencer_state._fields_ = [
     ('tr_avg_shift', uint16_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 90
+# /fast/lusee/coreloop/coreloop/core_loop.h: 93
 class struct_sequencer_program(Structure):
     pass
 
@@ -1157,12 +1165,13 @@ struct_sequencer_program._fields_ = [
     ('sequencer_repeat', uint16_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 99
+# /fast/lusee/coreloop/coreloop/core_loop.h: 102
 class struct_core_state_base(Structure):
     pass
 
 struct_core_state_base._pack_ = 1
 struct_core_state_base.__slots__ = [
+    'uC_time',
     'time_32',
     'time_16',
     'TVS_sensors',
@@ -1182,6 +1191,7 @@ struct_core_state_base.__slots__ = [
     'weight_current',
 ]
 struct_core_state_base._fields_ = [
+    ('uC_time', uint64_t),
     ('time_32', uint32_t),
     ('time_16', uint16_t),
     ('TVS_sensors', uint16_t * int(4)),
@@ -1201,7 +1211,7 @@ struct_core_state_base._fields_ = [
     ('weight_current', uint8_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 120
+# /fast/lusee/coreloop/coreloop/core_loop.h: 124
 class struct_delayed_cdi_sending(Structure):
     pass
 
@@ -1236,6 +1246,12 @@ struct_core_state.__slots__ = [
     'cdi_dispatch',
     'sequencer_enabled',
     'program',
+    'cmd_arg_high',
+    'cmd_arg_low',
+    'cmd_start',
+    'cmd_end',
+    'cmd_counter',
+    'dispatch_delay',
 ]
 struct_core_state._fields_ = [
     ('seq', struct_sequencer_state),
@@ -1243,9 +1259,15 @@ struct_core_state._fields_ = [
     ('cdi_dispatch', struct_delayed_cdi_sending),
     ('sequencer_enabled', c_bool),
     ('program', struct_sequencer_program),
+    ('cmd_arg_high', uint8_t * int(128)),
+    ('cmd_arg_low', uint8_t * int(128)),
+    ('cmd_start', uint16_t),
+    ('cmd_end', uint16_t),
+    ('cmd_counter', uint32_t),
+    ('dispatch_delay', uint16_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 143
+# /fast/lusee/coreloop/coreloop/core_loop.h: 151
 class struct_saved_core_state(Structure):
     pass
 
@@ -1261,7 +1283,7 @@ struct_saved_core_state._fields_ = [
     ('CRC', uint32_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 149
+# /fast/lusee/coreloop/coreloop/core_loop.h: 157
 class struct_startup_hello(Structure):
     pass
 
@@ -1287,7 +1309,7 @@ struct_startup_hello._fields_ = [
     ('time_16', uint16_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 160
+# /fast/lusee/coreloop/coreloop/core_loop.h: 168
 class struct_heartbeat(Structure):
     pass
 
@@ -1305,7 +1327,7 @@ struct_heartbeat._fields_ = [
     ('magic', c_char * int(6)),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 168
+# /fast/lusee/coreloop/coreloop/core_loop.h: 176
 class struct_meta_data(Structure):
     pass
 
@@ -1323,7 +1345,7 @@ struct_meta_data._fields_ = [
     ('base', struct_core_state_base),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 175
+# /fast/lusee/coreloop/coreloop/core_loop.h: 183
 class struct_housekeeping_data_base(Structure):
     pass
 
@@ -1341,7 +1363,7 @@ struct_housekeeping_data_base._fields_ = [
     ('housekeeping_type', uint16_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 182
+# /fast/lusee/coreloop/coreloop/core_loop.h: 190
 class struct_housekeeping_data_0(Structure):
     pass
 
@@ -1355,7 +1377,7 @@ struct_housekeeping_data_0._fields_ = [
     ('core_state', struct_core_state),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 187
+# /fast/lusee/coreloop/coreloop/core_loop.h: 195
 class struct_housekeeping_data_1(Structure):
     pass
 
@@ -1371,7 +1393,7 @@ struct_housekeeping_data_1._fields_ = [
     ('actual_gain', uint8_t * int(4)),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 193
+# /fast/lusee/coreloop/coreloop/core_loop.h: 201
 class struct_housekeeping_data_99(Structure):
     pass
 
@@ -1383,7 +1405,7 @@ struct_housekeeping_data_99._fields_ = [
     ('section_break', uint32_t),
 ]
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 199
+# /fast/lusee/coreloop/coreloop/core_loop.h: 207
 for _lib in _libs.values():
     try:
         state = (struct_core_state).in_dll(_lib, "state")
@@ -1391,7 +1413,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 200
+# /fast/lusee/coreloop/coreloop/core_loop.h: 208
 for _lib in _libs.values():
     try:
         soft_reset_flag = (c_bool).in_dll(_lib, "soft_reset_flag")
@@ -1399,7 +1421,7 @@ for _lib in _libs.values():
     except:
         pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 203
+# /fast/lusee/coreloop/coreloop/core_loop.h: 211
 for _lib in _libs.values():
     if not _lib.has("core_loop", "cdecl"):
         continue
@@ -1408,7 +1430,7 @@ for _lib in _libs.values():
     core_loop.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 206
+# /fast/lusee/coreloop/coreloop/core_loop.h: 214
 for _lib in _libs.values():
     if not _lib.has("process_cdi", "cdecl"):
         continue
@@ -1417,7 +1439,7 @@ for _lib in _libs.values():
     process_cdi.restype = c_bool
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 212
+# /fast/lusee/coreloop/coreloop/core_loop.h: 220
 for _lib in _libs.values():
     if not _lib.has("RFS_stop", "cdecl"):
         continue
@@ -1426,7 +1448,7 @@ for _lib in _libs.values():
     RFS_stop.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 213
+# /fast/lusee/coreloop/coreloop/core_loop.h: 221
 for _lib in _libs.values():
     if not _lib.has("RFS_start", "cdecl"):
         continue
@@ -1435,7 +1457,7 @@ for _lib in _libs.values():
     RFS_start.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 214
+# /fast/lusee/coreloop/coreloop/core_loop.h: 222
 for _lib in _libs.values():
     if not _lib.has("restart_spectrometer", "cdecl"):
         continue
@@ -1444,61 +1466,61 @@ for _lib in _libs.values():
     restart_spectrometer.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 217
+# /fast/lusee/coreloop/coreloop/core_loop.h: 225
 for _lib in _libs.values():
     if not _lib.has("get_Navg1", "cdecl"):
         continue
     get_Navg1 = _lib.get("get_Navg1", "cdecl")
-    get_Navg1.argtypes = [struct_core_state]
+    get_Navg1.argtypes = [POINTER(struct_core_state)]
     get_Navg1.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 218
+# /fast/lusee/coreloop/coreloop/core_loop.h: 226
 for _lib in _libs.values():
     if not _lib.has("get_Navg2", "cdecl"):
         continue
     get_Navg2 = _lib.get("get_Navg2", "cdecl")
-    get_Navg2.argtypes = [struct_core_state]
+    get_Navg2.argtypes = [POINTER(struct_core_state)]
     get_Navg2.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 219
+# /fast/lusee/coreloop/coreloop/core_loop.h: 227
 for _lib in _libs.values():
     if not _lib.has("get_Nfreq", "cdecl"):
         continue
     get_Nfreq = _lib.get("get_Nfreq", "cdecl")
-    get_Nfreq.argtypes = [struct_core_state]
+    get_Nfreq.argtypes = [POINTER(struct_core_state)]
     get_Nfreq.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 220
+# /fast/lusee/coreloop/coreloop/core_loop.h: 228
 for _lib in _libs.values():
     if not _lib.has("get_tr_avg", "cdecl"):
         continue
     get_tr_avg = _lib.get("get_tr_avg", "cdecl")
-    get_tr_avg.argtypes = [struct_core_state]
+    get_tr_avg.argtypes = [POINTER(struct_core_state)]
     get_tr_avg.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 221
+# /fast/lusee/coreloop/coreloop/core_loop.h: 229
 for _lib in _libs.values():
     if not _lib.has("get_gain_auto_max", "cdecl"):
         continue
     get_gain_auto_max = _lib.get("get_gain_auto_max", "cdecl")
-    get_gain_auto_max.argtypes = [struct_core_state, c_int]
+    get_gain_auto_max.argtypes = [POINTER(struct_core_state), c_int]
     get_gain_auto_max.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 222
+# /fast/lusee/coreloop/coreloop/core_loop.h: 230
 for _lib in _libs.values():
     if not _lib.has("get_tr_length", "cdecl"):
         continue
     get_tr_length = _lib.get("get_tr_length", "cdecl")
-    get_tr_length.argtypes = [struct_core_state]
+    get_tr_length.argtypes = [POINTER(struct_core_state)]
     get_tr_length.restype = uint32_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 226
+# /fast/lusee/coreloop/coreloop/core_loop.h: 234
 for _lib in _libs.values():
     if not _lib.has("set_route", "cdecl"):
         continue
@@ -1507,7 +1529,7 @@ for _lib in _libs.values():
     set_route.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 229
+# /fast/lusee/coreloop/coreloop/core_loop.h: 237
 for _lib in _libs.values():
     if not _lib.has("update_spec_gains", "cdecl"):
         continue
@@ -1516,7 +1538,7 @@ for _lib in _libs.values():
     update_spec_gains.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 232
+# /fast/lusee/coreloop/coreloop/core_loop.h: 240
 for _lib in _libs.values():
     if not _lib.has("trigger_ADC_stat", "cdecl"):
         continue
@@ -1525,7 +1547,7 @@ for _lib in _libs.values():
     trigger_ADC_stat.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 235
+# /fast/lusee/coreloop/coreloop/core_loop.h: 243
 for _lib in _libs.values():
     if not _lib.has("reset_errormasks", "cdecl"):
         continue
@@ -1534,7 +1556,7 @@ for _lib in _libs.values():
     reset_errormasks.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 238
+# /fast/lusee/coreloop/coreloop/core_loop.h: 246
 for _lib in _libs.values():
     if not _lib.has("update_time", "cdecl"):
         continue
@@ -1543,7 +1565,7 @@ for _lib in _libs.values():
     update_time.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 241
+# /fast/lusee/coreloop/coreloop/core_loop.h: 249
 for _lib in _libs.values():
     if not _lib.has("process_spectrometer", "cdecl"):
         continue
@@ -1552,7 +1574,7 @@ for _lib in _libs.values():
     process_spectrometer.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 244
+# /fast/lusee/coreloop/coreloop/core_loop.h: 252
 for _lib in _libs.values():
     if not _lib.has("transfer_to_cdi", "cdecl"):
         continue
@@ -1561,7 +1583,7 @@ for _lib in _libs.values():
     transfer_to_cdi.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 246
+# /fast/lusee/coreloop/coreloop/core_loop.h: 254
 for _lib in _libs.values():
     if not _lib.has("process_delayed_cdi_dispatch", "cdecl"):
         continue
@@ -1570,7 +1592,7 @@ for _lib in _libs.values():
     process_delayed_cdi_dispatch.restype = c_bool
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 249
+# /fast/lusee/coreloop/coreloop/core_loop.h: 257
 for _lib in _libs.values():
     if not _lib.has("process_gain_range", "cdecl"):
         continue
@@ -1579,7 +1601,7 @@ for _lib in _libs.values():
     process_gain_range.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 250
+# /fast/lusee/coreloop/coreloop/core_loop.h: 258
 for _lib in _libs.values():
     if not _lib.has("bitslice_control", "cdecl"):
         continue
@@ -1588,7 +1610,7 @@ for _lib in _libs.values():
     bitslice_control.restype = c_bool
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 253
+# /fast/lusee/coreloop/coreloop/core_loop.h: 261
 for _lib in _libs.values():
     if not _lib.has("set_spectrometer_to_sequencer", "cdecl"):
         continue
@@ -1597,7 +1619,7 @@ for _lib in _libs.values():
     set_spectrometer_to_sequencer.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 254
+# /fast/lusee/coreloop/coreloop/core_loop.h: 262
 for _lib in _libs.values():
     if not _lib.has("default_seq", "cdecl"):
         continue
@@ -1606,7 +1628,7 @@ for _lib in _libs.values():
     default_seq.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 255
+# /fast/lusee/coreloop/coreloop/core_loop.h: 263
 for _lib in _libs.values():
     if not _lib.has("advance_sequencer", "cdecl"):
         continue
@@ -1615,7 +1637,7 @@ for _lib in _libs.values():
     advance_sequencer.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 258
+# /fast/lusee/coreloop/coreloop/core_loop.h: 266
 for _lib in _libs.values():
     if not _lib.has("debug_helper", "cdecl"):
         continue
@@ -1624,7 +1646,7 @@ for _lib in _libs.values():
     debug_helper.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 259
+# /fast/lusee/coreloop/coreloop/core_loop.h: 267
 for _lib in _libs.values():
     if not _lib.has("cdi_not_implemented", "cdecl"):
         continue
@@ -1633,7 +1655,7 @@ for _lib in _libs.values():
     cdi_not_implemented.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 263
+# /fast/lusee/coreloop/coreloop/core_loop.h: 271
 for _lib in _libs.values():
     if not _lib.has("send_hello_packet", "cdecl"):
         continue
@@ -1642,7 +1664,7 @@ for _lib in _libs.values():
     send_hello_packet.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 264
+# /fast/lusee/coreloop/coreloop/core_loop.h: 272
 for _lib in _libs.values():
     if not _lib.has("process_hearbeat", "cdecl"):
         continue
@@ -1651,7 +1673,7 @@ for _lib in _libs.values():
     process_hearbeat.restype = c_bool
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 265
+# /fast/lusee/coreloop/coreloop/core_loop.h: 273
 for _lib in _libs.values():
     if not _lib.has("process_housekeeping", "cdecl"):
         continue
@@ -1660,7 +1682,7 @@ for _lib in _libs.values():
     process_housekeeping.restype = c_bool
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 276
+# /fast/lusee/coreloop/coreloop/core_loop.h: 284
 for _lib in _libs.values():
     if not _lib.has("mini_wait", "cdecl"):
         continue
@@ -1669,7 +1691,7 @@ for _lib in _libs.values():
     mini_wait.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 280
+# /fast/lusee/coreloop/coreloop/core_loop.h: 288
 for _lib in _libs.values():
     if not _lib.has("encode_12plus4", "cdecl"):
         continue
@@ -1678,7 +1700,7 @@ for _lib in _libs.values():
     encode_12plus4.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 281
+# /fast/lusee/coreloop/coreloop/core_loop.h: 289
 for _lib in _libs.values():
     if not _lib.has("decode_12plus4", "cdecl"):
         continue
@@ -1687,7 +1709,7 @@ for _lib in _libs.values():
     decode_12plus4.restype = c_int32
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 284
+# /fast/lusee/coreloop/coreloop/core_loop.h: 292
 for _lib in _libs.values():
     if not _lib.has("encode_10plus6", "cdecl"):
         continue
@@ -1696,7 +1718,7 @@ for _lib in _libs.values():
     encode_10plus6.restype = uint16_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 285
+# /fast/lusee/coreloop/coreloop/core_loop.h: 293
 for _lib in _libs.values():
     if not _lib.has("decode_10plus6", "cdecl"):
         continue
@@ -1705,7 +1727,7 @@ for _lib in _libs.values():
     decode_10plus6.restype = c_int32
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 291
+# /fast/lusee/coreloop/coreloop/core_loop.h: 299
 for _lib in _libs.values():
     if not _lib.has("encode_shared_lz_positive", "cdecl"):
         continue
@@ -1714,7 +1736,7 @@ for _lib in _libs.values():
     encode_shared_lz_positive.restype = c_int
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 292
+# /fast/lusee/coreloop/coreloop/core_loop.h: 300
 for _lib in _libs.values():
     if not _lib.has("decode_shared_lz_positive", "cdecl"):
         continue
@@ -1723,7 +1745,7 @@ for _lib in _libs.values():
     decode_shared_lz_positive.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 298
+# /fast/lusee/coreloop/coreloop/core_loop.h: 306
 for _lib in _libs.values():
     if not _lib.has("encode_shared_lz_signed", "cdecl"):
         continue
@@ -1732,7 +1754,7 @@ for _lib in _libs.values():
     encode_shared_lz_signed.restype = c_int
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 299
+# /fast/lusee/coreloop/coreloop/core_loop.h: 307
 for _lib in _libs.values():
     if not _lib.has("decode_shared_lz_signed", "cdecl"):
         continue
@@ -1741,7 +1763,7 @@ for _lib in _libs.values():
     decode_shared_lz_signed.restype = None
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 302
+# /fast/lusee/coreloop/coreloop/core_loop.h: 310
 for _lib in _libs.values():
     if not _lib.has("CRC", "cdecl"):
         continue
@@ -1750,93 +1772,105 @@ for _lib in _libs.values():
     CRC.restype = uint32_t
     break
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 9
+# /fast/lusee/coreloop/coreloop/core_loop.h: 6
 try:
-    VERSION_ID = 0x00000105
+    VERSION = '0.106-DEV'
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 20
+# /fast/lusee/coreloop/coreloop/core_loop.h: 9
+try:
+    VERSION_ID = 0x00000106
+except:
+    pass
+
+# /fast/lusee/coreloop/coreloop/core_loop.h: 20
 try:
     NSEQ_MAX = 32
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 21
+# /fast/lusee/coreloop/coreloop/core_loop.h: 21
 try:
     DISPATCH_DELAY = 6
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 22
+# /fast/lusee/coreloop/coreloop/core_loop.h: 22
 try:
-    RESETTLE_DELAY = 2
+    RESETTLE_DELAY = 5
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 23
+# /fast/lusee/coreloop/coreloop/core_loop.h: 23
 try:
     HEARTBEAT_DELAY = 1024
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 25
+# /fast/lusee/coreloop/coreloop/core_loop.h: 24
+try:
+    CMD_BUFFER_SIZE = 128
+except:
+    pass
+
+# /fast/lusee/coreloop/coreloop/core_loop.h: 27
 try:
     ADC_STAT_SAMPLES = 16000
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 27
+# /fast/lusee/coreloop/coreloop/core_loop.h: 29
 try:
     MAX_STATE_SLOTS = 64
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 29
+# /fast/lusee/coreloop/coreloop/core_loop.h: 31
 try:
     PAGES_PER_SLOT = 256
 except:
     pass
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 273
+# /fast/lusee/coreloop/coreloop/core_loop.h: 281
 def MAX(x, y):
     return (x > y) and x or y
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 274
+# /fast/lusee/coreloop/coreloop/core_loop.h: 282
 def MIN(x, y):
     return (x < y) and x or y
 
-# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 275
+# /fast/lusee/coreloop/coreloop/core_loop.h: 283
 def IS_NEG(x):
     return (x < 0) and 1 or 0
 
-core_state = struct_core_state# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 134
+core_state = struct_core_state# /fast/lusee/coreloop/coreloop/core_loop.h: 138
 
-route_state = struct_route_state# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 66
+route_state = struct_route_state# /fast/lusee/coreloop/coreloop/core_loop.h: 69
 
-sequencer_state = struct_sequencer_state# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 72
+sequencer_state = struct_sequencer_state# /fast/lusee/coreloop/coreloop/core_loop.h: 75
 
-sequencer_program = struct_sequencer_program# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 90
+sequencer_program = struct_sequencer_program# /fast/lusee/coreloop/coreloop/core_loop.h: 93
 
-core_state_base = struct_core_state_base# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 99
+core_state_base = struct_core_state_base# /fast/lusee/coreloop/coreloop/core_loop.h: 102
 
-delayed_cdi_sending = struct_delayed_cdi_sending# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 120
+delayed_cdi_sending = struct_delayed_cdi_sending# /fast/lusee/coreloop/coreloop/core_loop.h: 124
 
-saved_core_state = struct_saved_core_state# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 143
+saved_core_state = struct_saved_core_state# /fast/lusee/coreloop/coreloop/core_loop.h: 151
 
-startup_hello = struct_startup_hello# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 149
+startup_hello = struct_startup_hello# /fast/lusee/coreloop/coreloop/core_loop.h: 157
 
-heartbeat = struct_heartbeat# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 160
+heartbeat = struct_heartbeat# /fast/lusee/coreloop/coreloop/core_loop.h: 168
 
-meta_data = struct_meta_data# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 168
+meta_data = struct_meta_data# /fast/lusee/coreloop/coreloop/core_loop.h: 176
 
-housekeeping_data_base = struct_housekeeping_data_base# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 175
+housekeeping_data_base = struct_housekeeping_data_base# /fast/lusee/coreloop/coreloop/core_loop.h: 183
 
-housekeeping_data_0 = struct_housekeeping_data_0# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 182
+housekeeping_data_0 = struct_housekeeping_data_0# /fast/lusee/coreloop/coreloop/core_loop.h: 190
 
-housekeeping_data_1 = struct_housekeeping_data_1# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 187
+housekeeping_data_1 = struct_housekeeping_data_1# /fast/lusee/coreloop/coreloop/core_loop.h: 195
 
-housekeeping_data_99 = struct_housekeeping_data_99# /home/anze/Dropbox/work/lusee/coreloop/coreloop/core_loop.h: 193
+housekeeping_data_99 = struct_housekeeping_data_99# /fast/lusee/coreloop/coreloop/core_loop.h: 201
 
 # No inserted files