+++ src/spectrometer_interface.c
@@ -313,5 +313,7 @@ void spec_set_ADC_toggle_pattern() {}
 void spec_set_ADC_load_custom_pattern() {}
 void spec_set_ADC_custom_pattern() {}
 
+void spec_set_fw_cdi_delay(uint32_t delay) {}
+
  uint32_t spec_read_uC_register(uint8_t num) {return boot_registers[num];}
  void spec_write_uC_register(uint8_t num, uint32_t value) {boot_registers[num] = value;}